requirements.txt
python-dotenv==1.0.0
requests==2.31.0
moviepy==1.0.3
gTTS==2.3.2
instabot==0.117.0
Pillow>=9.5.0
numpy==1.24.3
